implemented persistent section collapsed state

+++ app/src/main/java/io/github/bradpatras/todometer/core/domain/AppPreferences.kt
@@ -1,5 +1,5 @@
 package io.github.bradpatras.todometer.core.domain
 
 data class AppPreferences(
-    val collapsedSectionIds: List<Int>
+    val collapsedSectionIds: List<Long>
 )
